second commit

+++ src/main/java/TDD/SimpleCalculator.java
@@ -2,7 +2,7 @@ package TDD;
 
 public class SimpleCalculator {
 
-    public int add(int num1, int num2) {
+    public long add(long num1, long num2) {
         return num1 + num2;
     }
 }
