+++ src/test/java/SimpleCalculatorTest.java
@@ -7,7 +7,7 @@ public class SimpleCalculatorTest {
     @Test
     public void testDisplayStudentName() {
         SimpleCalculator simpleCalculator = new SimpleCalculator();
-        int sumOfTwoNumbers = simpleCalculator.add(10, 20);
+        long sumOfTwoNumbers = simpleCalculator.add(10, 20);
 
         Assert.assertEquals(30, sumOfTwoNumbers);
     }
